requirements.txt
flask==3.0.3
flask-wtf==1.2.2
flask-limiter==3.7.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
flask-sqlalchemy==3.1.1
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.51.2
reportlab==4.2.2
